requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
Django==3.2.5
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-registration==2.4.1
django-tinymce==3.3.0
django-utils-six==2.0
Pillow==8.2.0
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
